+++ src/types/pallet-events/index.ts
@@ -1,0 +1,3 @@
+export * from './shielded-pool';
+export * from './zk-verifier';
+export * from './account-mapping';